refactor: move filtering of sockets out of router
